requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
deprecation==2.1.0
exceptiongroup==1.2.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gotrue==2.5.4
gspread==6.1.2
gspread-dataframe==4.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
lxml==5.2.2
more-itertools==10.3.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
postgrest==0.16.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.1
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
zeep==4.2.1
